# Service/src/main/webapp/dist/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an API for manipulating subjects at faculty",
        "version": "1.0.0",
        "title": "Registar predmeta",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "crncevic.petar1@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "host": "localhost:8080",
    "basePath": "/Service-1.0-SNAPSHOT/api",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/predmet/{id}": {
            "get": {
                "tags": [
                    "predmet"
                ],
                "summary": "Find predmet by id",
                "description": "Returns a single predmet",
                "operationId": "getById",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of predmet to return",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "204": {
                        "description": "No content"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "put": {
                "tags": [
                    "predmet"
                ],
                "summary": "Updates predmet with form data",
                "description": "",
                "operationId": "update",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "predmetDTO",
                        "in": "formData",
                        "description": "Updated data for predmet",
                        "required": true,
                        "type": "PredmetDTO"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of predmet for update",
                        "required": true,
                        "type": "int32"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "predmet succesfully updated"
                    },
                    "400": {
                        "description": "Bad request. Model constraints are not satisfied"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "predmet"
                ],
                "summary": "Deletes predmet with all udzbenik na predmetnu , nastavnik na predmetu and predmet na studijskom programu!",
                "description": "",
                "operationId": "delete",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Predmet id to delete",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Predmet succesfully deleted"
                    },
                    "400": {
                        "description": "Bad request. Model constraints are not satisfied"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
...
        "/tematska-celina/predmet/{predmetId}": {
            "get": {
                "tags": [
                    "tematska-celina"
                ],
                "summary": "Find all tematska celina by predmetId",
                "description": "Returns a list of tematska celina for predmet",
                "operationId": "getByPredmetId",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "predmetId",
                        "in": "path",
                        "description": "ID of predmet to return tematske-celine",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "204": {
                        "description": "No content"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
...
        "/predmet-na-std-programu/{stdProgramId}": {
            "get": {
                "tags": [
                    "predmet na studijskom programu"
                ],
                "summary": "Find predmet by stdProgramId",
                "description": "Returns a list of subjects at std program",
                "operationId": "getByStdProgramId",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "stdProgramId",
                        "in": "path",
                        "description": "ID of std program for his subjects to return",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "204": {
                        "description": "No content"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "predmet-na-std-programu/{stdProgramId}/{predmetId}":{
            "get": {
                "tags": [
                    "predmet na studijskom programu"
                ],
                "summary": "Find predmet by stdProgramId and predmetId",
                "description": "Returns subject at std program",
                "operationId": "getById",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "stdProgramId",
                        "in": "path",
                        "description": "ID of std program for his subjects to return",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    },
                    {
                        "name": "predmetId",
                        "in": "path",
                        "description": "ID of subject",
                        "required": true,
                        "type": "integer",
                        "format": "int32"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "204": {
                        "description": "No content"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
        }
    }
}
}